Set up api to add players
