--- game/createPlayer.js
@@ -1,16 +1,0 @@
-const Player = require('../models/player');
-
-const createPlayer = (req, res) => {
-
-    const player = new Player({
-        name: req.body.name,
-        points: 0,
-        type: "N"
-    });
-
-    player.save()
-        .then(doc=>res.send(doc))
-        .catch(err=> console.log(error))
-};
-
-module.exports = {createPlayer};
