+++ game/player.js
@@ -1,0 +1,16 @@
+const Player = require('../models/room');
+
+const createPlayer = (req, res) => {
+
+    const player = new Player({
+        name: req.body.name,
+        points: 0,
+        type: "N"
+    });
+
+    player.save()
+        .then(doc=>res.send(doc))
+        .catch(err=> console.log(err))
+};
+
+module.exports = {createPlayer};
