+++ models/room.js
@@ -32,3 +32,4 @@ const roomSchema = new Schema({
 });
 
 module.exports = mongoose.model('Room', roomSchema);
+module.exports = mongoose.model('Player', playerSchema);
