+++ package.json
@@ -20,6 +20,7 @@
   "homepage": "https://github.com/nkalia1802/GWAP-backend#readme",
   "dependencies": {
     "body-parser": "^1.19.0",
+    "cors": "^2.8.5",
     "express": "^4.17.1",
     "mongodb": "^3.6.6",
     "mongoose": "^5.12.8"
